sms-code/requirements.txt
# requirements.txt

APScheduler==3.10.4
asarPy==1.0.1
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
crispy-bootstrap4==2024.1
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
django-admin-adminlte==1.0.5
django-appconf==1.0.6
django-bootstrap4==24.1
django-braces==1.15.0
django-crispy-forms==2.1
django-crispy-formset-modal==0.6
django-extra-views==0.14.0
django-filter==24.1
django-phonenumber-field==7.3.0
django-postgres-extra==2.0.9rc10
django-redis==5.4.0
django-report-builder @ https://codeload.github.com/TuringMachinaTest/django-report-builder/zip/refs/heads/master#sha256=7ff7fe229dd20a0ff7bbf853f4ed52a2cfb427039be144daaf082f21c93a80e8
django-select2==8.1.2
django-simple-history==3.5.0
django-tables2==2.7.0
django-view-breadcrumbs==2.5.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
openpyxl==3.1.2
packaging==24.0
phonenumbers==8.13.32
phonenumberslite==8.13.32
priority==1.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
str2bool==1.1
tablib==3.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
whitenoise==6.6.0
zope.interface==6.2
